0.0.4.7 * advanced file dialog returned * Community path detection (Rikoooo's method) * AIR import process simplified * Notice about how to adjust center of gravity after fuselage length insertion * ImageTools texture conversion button moved to 1st position * AIR tables updated (Taz5150's data) * Locked files error handlers * FSX gauge expressions parsing improved

+++ fs2020LegacyImporter/Extract.cs
@@ -13,7 +13,7 @@ namespace msfsLegacyImporter
         {
         }
 
-        public void Run(string TEMP_FILE, string EXE_PATH)
+        public void Run(string TEMP_FILE, string EXE_PATH, string targetDir)
         {
             BackgroundWorker bgw = new BackgroundWorker
             {
@@ -25,20 +25,24 @@ namespace msfsLegacyImporter
                     BackgroundWorker bw = o as BackgroundWorker;
                     FastZip fastZip = new FastZip();
                     Console.WriteLine("Unzipping");
-                    fastZip.ExtractZip(TEMP_FILE, AppDomain.CurrentDomain.BaseDirectory + "\\", FastZip.Overwrite.Always, null, @"-(ICSharpCode.SharpZipLib.dll)", null, false);
+                    fastZip.ExtractZip(TEMP_FILE, targetDir, FastZip.Overwrite.Always, null, @"-(ICSharpCode.SharpZipLib.dll)", null, false);
                 });
 
             bgw.RunWorkerCompleted += new RunWorkerCompletedEventHandler(
             delegate (object o, RunWorkerCompletedEventArgs args) {
                 //((MainWindow)System.Windows.Application.Current.MainWindow).setUpdateReady();
-                if (File.Exists(EXE_PATH))
+                if (!String.IsNullOrEmpty(EXE_PATH))
                 {
-                    Process.Start(EXE_PATH);
-                    Environment.Exit(0);
-                } else
-                {
-                    MessageBox.Show("Update failed, but you can extract temp.zip manually");
-                    File.Move(EXE_PATH + ".BAK", EXE_PATH);
+                    if (File.Exists(EXE_PATH))
+                    {
+                        Process.Start(EXE_PATH);
+                        Environment.Exit(0);
+                    }
+                    else
+                    {
+                        MessageBox.Show("Update failed, but you can extract temp.zip manually");
+                        File.Move(EXE_PATH + ".BAK", EXE_PATH);
+                    }
                 }
             });
